Refactor flytepropeller to allow external start
